refactor: create `asset-config.ts`
